Phase 4: Decompose characters.py into armor, clothing, and appearance mixins
Extract 3 mixin classes from the 2,117-line Character monolith:

- ArmorMixin (475 lines): take_damage, armor damage reduction, plate carrier
  expansion, armor effectiveness, armor degradation
- ClothingMixin (387 lines): wear/remove items, layer conflict detection,
  coverage queries, coverage map builder
- AppearanceMixin (631 lines): longdesc appearance, body descriptions,
  paragraph formatting, pronoun resolution, return_appearance hook

Character class inherits from all three mixins via MRO, keeping its import
path (typeclasses.characters.Character) unchanged. Zero consumer files
modified — all 15+ files using isinstance(), Character.objects, or string
typeclass references continue working without changes.

characters.py: 2,117 → 1,008 lines (52.4% reduction)

Fixes #56
